# internal/domains/blog.go
package domains

import (
	"time"
)

// Blog represents a blog post.
type Blog struct {
	ID        string     `db:"id"`
	Title     string     `db:"title"`
	Content   string     `db:"content"`
	CreatedAt *time.Time `db:"created_at"`
	UpdatedAt *time.Time `db:"updated_at"`
	DeletedAt *time.Time `db:"deleted_at"`
}